Update BarUpDown.cs
Fixed:
Was taking extra trades/drawing trade entry arrows since the the entry condition for taking Long/Short was getting triggered even when we were already in a trade. Added condition for market position to prevent this
Moved the Long/Short entry below the switch statement. Need to consider moving the code within the switch statement...
Added a  few extra print statements
uncommented the moving of SL to Break Even... so far not able to test.